Drop python 3.6 test since requirements require python >= 3.7
