+++ gpvolve/__init__.py
@@ -1,10 +1,6 @@
 # Import the main module in this package
 from gpvolve.__version__ import __version__
-from gpvolve.base import *
-from gpvolve.check import *
-from gpvolve.flux import *
-from gpvolve.matrix import *
-from gpvolve.paths import *
+from gpvolve.base import build_transition_matrix, apply_fitness_function
 from gpvolve import simulate
 from gpvolve import utils
 from gpvolve import pyplot
@@ -13,4 +9,4 @@ from gpvolve import markov
 from gpvolve import analysis
 from gpvolve import phenotype_to_fitness
 # Import fixation models with simpler names
-import gpvolve.markov.utils._generate_tmatrix.fixation as fixation_models
+# import gpvolve.markov.base._generate_tmatrix.fixation as fixation_models
