+++ gpvolve/markov/__init__.py
@@ -1,1 +1,1 @@
-from gpvolve.markov import utils
+from . import base
