+++ gpvolve/simulate/__init__.py
@@ -1,2 +1,1 @@
-from gpvolve.simulate import *
-from gpvolve.simulate.wright_fisher import wf_engine
+from .simulate import simulate
