+++ gpvolve/utils/__init__.py
@@ -1,3 +1,5 @@
 from gpvolve.utils.flatten_neighbors import flatten_neighbors
 from gpvolve.utils.check_neighbor_connectivity import check_neighbor_connectivity
 from gpvolve.utils.base import *
+from gpvolve.utils.peaks import *
+from gpvolve.utils.neutral_networks import *
