+++ tests/markov/utils/generate_tmatrix/test_fixation_models.py
@@ -1,7 +1,7 @@
 import pytest
 
-import gpvolve.markov.utils._generate_tmatrix.generate_tmatrix_cython as cy
-import gpvolve.markov.utils._generate_tmatrix.generate_tmatrix_python as py
+import gpvolve.markov.base._generate_tmatrix.generate_tmatrix_cython as cy
+import gpvolve.markov.base._generate_tmatrix.generate_tmatrix_python as py
 
 import numpy as np
