+++ tox.ini
@@ -1,17 +1,10 @@
 [tox]
-envlist = py27, py34, py35, py36, flake8
+envlist = py34, py35, flake8
 
 [travis]
 python =
-    3.6: py36
     3.5: py35
     3.4: py34
-    2.7: py27
-
-[testenv:flake8]
-basepython = python
-deps = flake8
-commands = flake8 gpvolve
 
 [testenv]
 setenv =
